requirements.txt
Django==1.11.3
django-filebrowser-no-grappelli==3.7.2
django-hitcount==1.3.0
django-tinymce4-lite==1.4.2
olefile==0.44
Pillow==4.2.1
pytz==2017.2
